pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "pillow>=11.3.0",
    "pytz>=2025.2",
]
